requirements.txt
matplotlib==3.8.2
numpy==1.24.1
opencv_python_headless==4.9.0.80
pandas==2.2.2
Pillow==10.4.0
pydicom==2.4.4
pytorch_lightning==2.1.3
scipy==1.14.0
seaborn==0.13.2
skimage
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.66.1
